Making fallback to Cookies configurable
When private browsing is used then local & session storage is not available and using it throws, and there is no way to check if it is supported. Last few check-ins added a fallback to cookies to provide any storage in a private browsing mode. This check-in provides a way to configure whether this module should fallback to cookies or not.